requirements.txt
# Core Framework
flask==2.3.3
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.4.2

# Cryptography Libraries
cryptography==41.0.4
pycryptodome==3.19.0
pyopenssl==23.2.0

# Post-Quantum Cryptography
# Note: liboqs-python cần được cài đặt riêng
# pip install git+https://github.com/open-quantum-safe/liboqs-python.git

# Identity-Based Encryption
# Note: Charm-Crypto cần được build từ source
# Xem hướng dẫn tại: https://jhuisi.github.io/charm/install_source.html

# Database
sqlalchemy==2.0.21
psycopg2-binary==2.9.7
alembic==1.12.0
redis==5.0.0

# Security & Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.0.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7
colorlog==6.7.0

# API & Web
requests==2.31.0
httpx==0.25.0
python-dateutil==2.8.2
pytz==2023.3

# Monitoring & Logging
prometheus-client==0.17.1
python-json-logger==2.0.7

# Testing
pytest==7.4.2
pytest-asyncio==0.21.1
pytest-cov==4.1.0
faker==19.6.2

# Development
black==23.9.1
flake8==6.1.0
mypy==1.5.1
pre-commit==3.4.0pqcrypto==0.1.0
cryptography==41.0.7
hvac==1.2.1
